:construction: Add wall detection for alien hord
Everytime the `updateEntities` is updating the first enemy alive from
the hord it checks if the hord is touching or beyond the left limit or
the right limit, if that happens it will reverse the velocity on x
component on all enemies from the hord that are on the Entity array from
element 2 to `FIRST_IDX_BULLETS`-1.
The first approach I tried was to check the left wall on the first enemy update and the wall on the right when updating the last enemy alive, but that was incorrect given the fact that when the hord bounce on the right side the algorithm notices it only on the last enemy update, in that case all the enemies was beyond the limit on the right side.
- The aliens move and bounce on the walls;
- It's a little hacky right now but can be improved;
- Also changes the size of the player ship, it was too wide.