:construction: More decoupling
As `detectCollisions` seems to do something that can be perceived as an
updating on the entities, given the fact that if a bullet collide with
another entity both will die, sounds correct to put that inside
`updateEntities`. In addition the regular check of
`gameState.exitWindowRequested` was moved from `mainloop` to
`updateEntities`.
Moreover, `drawEntities` was changed to `drawGame` because now it draws
the exit confirmation message.